Change dt from uint to double
* Creates less choppy animations

+++ Ongkingco-PA3/Ongkingco-PA3/graphics.cpp
@@ -92,7 +92,7 @@ bool Graphics::Initialize(int width, int height)
   return true;
 }
 
-void Graphics::Update(unsigned int dt, glm::vec3 speed)
+void Graphics::Update(double dt, glm::vec3 speed)
 {
   // Update the cube
   m_triangle->Update(dt);
